# week6/DL-07-Agentic-AI-System-II/02_api_backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertsBody": {
        "additionalProperties": false,
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/AlertChannel"
            },
            "maxItems": 10,
            "title": "Channels",
            "type": "array"
          },
          "consent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent At"
          },
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "title": "AlertsBody",
        "type": "object"
      },
...
      "AlertsView": {
        "additionalProperties": false,
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/AlertChannel"
            },
            "title": "Channels",
            "type": "array"
          },
          "consent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent At"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled",
          "consent_at",
          "channels"
        ],
        "title": "AlertsView",
        "type": "object"
      },
      "AssessmentCreate": {
        "additionalProperties": false,
        "properties": {
          "language": {
            "anyOf": [
              {
                "maxLength": 35,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "mode": {
            "$ref": "#/components/schemas/RequestMode",
            "default": "auto"
          }
        },
        "title": "AssessmentCreate",
        "type": "object"
      },
...
      "FieldErrorItem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "code"
        ],
        "title": "FieldErrorItem",
        "type": "object"
      },
...
      "JobAccepted": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "estimated_seconds": {
            "default": 20,
            "title": "Estimated Seconds",
            "type": "integer"
          },
          "events_url": {
            "title": "Events Url",
            "type": "string"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "recommendation_id": {
            "format": "uuid",
            "title": "Recommendation Id",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "default": "queued",
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "recommendation_id",
          "conversation_id",
          "events_url",
          "status_url"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
...
      "Location": {
        "additionalProperties": false,
        "properties": {
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "place_id": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place Id"
          }
        },
        "required": [
          "lat",
          "lon"
        ],
        "title": "Location",
        "type": "object"
      },
...
      "ProblemResponse": {
        "additionalProperties": false,
        "description": "RFC 9457 Problem Details \u2014 the shape of every error response (app/api/problem.py).\n\nFastAPI documents its own `HTTPValidationError` for 422 by default; `custom_openapi`\nin app/main.py replaces every generated 422 (and adds this for other error statuses\ndeclared per-route) with this schema so the exported contract matches what the API\nactually sends.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldErrorItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "instance",
          "request_id",
          "correlation_id",
          "errors"
        ],
        "title": "ProblemResponse",
        "type": "object"
      },
...
      "RecommendationPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "RecommendationPage",
        "type": "object"
      },
      "RecommendationResponse": {
        "additionalProperties": false,
        "properties": {
          "clarification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Clarification"
              },
              {
                "type": "null"
              }
            ]
          },
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_freshness": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataFreshness"
              },
              {
                "type": "null"
              }
            ]
          },
          "disclaimer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disclaimer"
          },
          "emergency_instructions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmergencyInstructions"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "hazards": {
            "items": {
              "$ref": "#/components/schemas/Hazard"
            },
            "title": "Hazards",
            "type": "array"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendation_id": {
            "format": "uuid",
            "title": "Recommendation Id",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "risk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Risk"
              },
              {
                "type": "null"
              }
            ]
          },
          "routes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Routes"
              },
              {
                "type": "null"
              }
            ]
          },
          "service_status": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ServiceState"
            },
            "title": "Service Status",
            "type": "object"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceItem"
            },
            "title": "Sources",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/RecommendationStatus"
          },
          "valid_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "versions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Versions"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningItem"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "recommendation_id",
          "conversation_id",
          "request_id",
          "status",
          "created_at"
        ],
        "title": "RecommendationResponse",
        "type": "object"
      },
...
      "RequestMode": {
        "enum": [
          "auto",
          "sync",
          "async"
        ],
        "title": "RequestMode",
        "type": "string"
      },
...
      "TripCreate": {
        "additionalProperties": false,
        "properties": {
          "alerts": {
            "$ref": "#/components/schemas/AlertsBody"
          },
          "departure_time": {
            "format": "date-time",
            "title": "Departure Time",
            "type": "string"
          },
          "destination": {
            "$ref": "#/components/schemas/Location"
          },
          "name": {
            "maxLength": 1000,
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/Location"
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          },
          "timezone": {
            "maxLength": 64,
            "title": "Timezone",
            "type": "string"
          },
          "waypoints": {
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "maxItems": 50,
            "title": "Waypoints",
            "type": "array"
          }
        },
        "required": [
          "name",
          "origin",
          "destination",
          "departure_time",
          "timezone"
        ],
        "title": "TripCreate",
        "type": "object"
      },
      "TripPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TripResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "TripPage",
        "type": "object"
      },
      "TripPatch": {
        "additionalProperties": false,
        "description": "JSON Merge Patch (RFC 7396) of a trip; `null` removes or is rejected (D-60).",
        "properties": {
          "alerts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlertsPatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "departure_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Time"
          },
          "destination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "origin": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "preferences": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreferenceChanges"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "timezone": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "waypoints": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Location"
                },
                "maxItems": 50,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waypoints"
          }
        },
        "title": "TripPatch",
        "type": "object"
      },
      "TripResponse": {
        "additionalProperties": false,
        "properties": {
          "alerts": {
            "$ref": "#/components/schemas/AlertsView"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "departure_time": {
            "format": "date-time",
            "title": "Departure Time",
            "type": "string"
          },
          "destination": {
            "$ref": "#/components/schemas/Location"
          },
          "last_assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LastAssessment"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/Location"
          },
          "preferences": {
            "$ref": "#/components/schemas/Preferences"
          },
          "status": {
            "$ref": "#/components/schemas/TripStatus"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "trip_id": {
            "format": "uuid",
            "title": "Trip Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "waypoints": {
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "title": "Waypoints",
            "type": "array"
          }
        },
        "required": [
          "trip_id",
          "name",
          "origin",
          "destination",
          "waypoints",
          "departure_time",
          "timezone",
          "preferences",
          "alerts",
          "status",
          "last_assessment",
          "created_at",
          "updated_at"
        ],
        "title": "TripResponse",
        "type": "object"
      },
      "TripStatus": {
        "enum": [
          "PLANNED",
          "ACTIVE",
          "COMPLETED",
          "CANCELLED"
        ],
        "title": "TripStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Travel Safety & Advisory API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/trips": {
      "get": {
        "operationId": "list_trips_v1_trips_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TripStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Saved trips, latest departure first",
        "tags": [
          "trips"
        ]
      },
      "post": {
        "operationId": "create_trip_v1_trips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Save a trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/trips/{trip_id}": {
      "delete": {
        "operationId": "delete_trip_v1_trips__trip_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Delete a trip (its assessments stay in the history)",
        "tags": [
          "trips"
        ]
      },
      "get": {
        "operationId": "get_trip_v1_trips__trip_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Get a trip",
        "tags": [
          "trips"
        ]
      },
      "patch": {
        "operationId": "update_trip_v1_trips__trip_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripPatch"
              }
            },
            "application/merge-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/TripPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Change a trip (JSON Merge Patch)",
        "tags": [
          "trips"
        ]
      }
    },
    "/v1/trips/{trip_id}/assessments": {
      "get": {
        "operationId": "list_assessments_v1_trips__trip_id__assessments_get",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Assessments of the trip, newest first",
        "tags": [
          "trips"
        ]
      },
      "post": {
        "operationId": "assess_trip_v1_trips__trip_id__assessments_post",
        "parameters": [
          {
            "in": "path",
            "name": "trip_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Trip Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "accept-language",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Accept-Language"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AssessmentCreate"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Still running; follow the job"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemResponse"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Assess the trip now",
        "tags": [
          "trips"
        ]
      }
    }
  }
}